Prepare for open-sourcing
Restructured and changed the code to be more readable (we were on a time limit,
sorry!) by making it comply with K&R style. Also, removed useless commented out
cout statements and commented the code to make it more understandable.

NOTE: I did NOT change any functionaliy of the code, this is just for opening
the source to everybody to see and evaluate, especially everybody preparing for
the next MegaminerAI event!