Total AI cleanup
We didnt want to change to much within the original repo of our megaminer AI code in terms of cleanup and documentation in order to maintain the code we wrote while under a strict timelimit. But in order for this code to be used as an example in presentation in the future we started this repo in order to redocument and clean up the AI code, while maintaining the original logic.